pyproject.toml
[tool.poetry]
name = "repl-nix-deploymentbuddy"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.44.0"
python-telegram-bot = "^21.5"
anthropic = "^0.34.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
